requirements.txt
python-dotenv==0.10.3
requests==2.22.0
instabot==0.60.0
Pillow==6.1.0